WebPage/requirements.txt

algoliasearch-django>=3.0,<4.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
requests==2.31.0